+++ src/core/validator.ts
@@ -1,0 +1,3 @@
+export function isOptions(arg: unknown): arg is { flatten?: boolean } {
+  return typeof arg === 'object' && arg !== null && 'flatten' in arg;
+}
